server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.3.0",
    "beanie>=1.30.0",
    "fastapi>=0.115.13",
    "fastembed>=0.7.1",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.26",
    "langchain-community>=0.3.26",
    "langchain-google-genai>=2.0.10",
    "langchain-qdrant>=0.2.0",
    "motor>=3.7.1",
    "pydantic[email]>=2.11.7",
    "pyjwt>=2.10.1",
    "pymongo>=4.13.2",
    "pymupdf>=1.26.1",
    "python-dotenv>=1.1.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "qdrant-client[fastembed]>=1.14.3",
    "uuid>=1.30",
    "uvicorn>=0.34.3",
]
